Refactor: user.entity.ts 위치 변경
user.entity.ts 위치를 entity 디렉토리 하위로 변경, import 부분 수정
