+++ be/src/auth/entity/user.entity.ts
